Update DESCRIPTION.md (#29)
* Update DESCRIPTION.md

As a beginner, I've found this nugget of information essential. From the perspective of someone coming onto the site and working from beginning to end, leaving out information on where to find these numbers (other than looking at the expected valeus displayed in the strace) makes it very difficult. 

I found it referenced in this video: https://www.youtube.com/watch?v=AX4FOHwTFwo

Feel free to edit the text/spelling/tone, but as long as the core information is present, I think this is a necessary addition

* Update DESCRIPTION.md

---------

Co-authored-by: Yan Shoshitaishvili <yans@yancomm.net>